requirements.txt
dataclasses==0.6
Django==2.2.3
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
mypy==0.711
mypy-extensions==0.4.1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
django-cors-headers==3.0.2
gunicorn
django-heroku
